module layout (not ready) and made stealth code readable (...)
